% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestModel-package.R
\docType{package}
\name{BestModel}
\alias{BestModel}
\alias{BestModel-package}
\title{BestModel: Visualization for Machine Learning Metrics}
\description{
Want to easily compare Machine Learning models?  Want to just summarize them in a pretty graph
without having to go to all the work?  This package consolidates this by allowing you to use
your pre-built models to make easy graphics or to build all those models for you.
}
