% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{BuildModels}
\alias{BuildModels}
\title{A helper function used by the GetModelComparison function. It will take the given inputs
and build the models for the ModelComparison object.}
\usage{
BuildModels(training.data, training.classes, trctrl, tune.length,
  multi.class, build.flags, force.prepared = F)
}
\arguments{
\item{training.data}{The training data to build the models.}

\item{training.classes}{The labels for the training data.}

\item{trctrl}{The trctrl used for the caret package to train the models.}

\item{tune.length}{The amound of time to tune to the model. Defaults to 1.}

\item{multi.class}{A boolean value of whether the model has multiple classification outputs}

\item{build.flags}{A vector of flags that decide what models to build.}

\item{force.prepared}{A boolean value describing whether the dataset was forced prepared into
one hot encoding for training.  Only applicable with categorical data}
}
\value{
A vector of the models created.
}
\description{
A helper function used by the GetModelComparison function. It will take the given inputs
and build the models for the ModelComparison object.
}
\examples{
# This function is used by GetModelComparison. It's usage is as follows:
modelVec = BuildModels(training.data, training.classes, trctrl,
tune.length, multi.class, build.flags, force.prepared = forced.prepared)

}
\keyword{internal}
