% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{CreateModelComparison}
\alias{CreateModelComparison}
\title{A helper function used by GetModelComparison and ModelComparison to create the
ModelComparison object. It gives the object it's class and assigns it it's members}
\usage{
CreateModelComparison(model.list, multi.class, force.prepped,
  diff.names = NULL)
}
\arguments{
\item{model.list}{a list of models that are already trained.  The name of the model.list should
be accurate to the names of the models, unless other names are given in diff.names}

\item{multi.class}{a boolean value of whether the classification output is more than two outputs}

\item{force.prepped}{a boolean value describing whether the dataset was forced prepared into
one hot encoding for training.  Only applicable with categorical data}

\item{diff.names}{a vector or list containing the names of the model.list models, in that
respective order.  This is optional if the names are already attatched to the model.list
 and will default to NULL.}
}
\value{
the fully created ModelComparison object
}
\description{
A helper function used by GetModelComparison and ModelComparison to create the
ModelComparison object. It gives the object it's class and assigns it it's members
}
\examples{
# This function is used by GetModelComparison and ModelComparison to create the
ModelComparison object. The follow code is what is used by ModelComparison:
CreateModelComparison(model.list, multi.class, F, names(model.list))

}
\keyword{internal}
