% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{CreateROCPlot}
\alias{CreateROCPlot}
\title{A helper function used by the plot.ModelComparison.  This function gathers the data needed
for the ROC lines and then plots the chart.}
\usage{
CreateROCPlot(object, pred.basic, labels)
}
\arguments{
\item{object}{The ModelComparison object.}

\item{pred.basic}{The predictions from the models on the testing data.}

\item{labels}{The labels for the testing data.}
}
\description{
A helper function used by the plot.ModelComparison.  This function gathers the data needed
for the ROC lines and then plots the chart.
}
\examples{
# This function is used by plot.ModelComparison. It's usage is as follows:
 CreateROCPlot(object, pred.basic, labels)
}
\keyword{internal}
