% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembling.R
\name{Ensemble}
\alias{Ensemble}
\title{Ensemble Creation}
\usage{
Ensemble(model.list, voting.type, weights = "none", test.set = "none",
  train.type = "Balanced Accuracy")
}
\arguments{
\item{model.list}{The models to be used in the Ensemble}

\item{voting.type}{The voting type of the Ensemble (average vote, majority vote,
or majority weight)}

\item{weights}{A vector of length as long as the model.list containing numeric percent values
c(.76, .56, etc.) to be used for a weighted vote Ensemble. For auto calculation, leave
this blank and provide a test set.  If the Ensemble is not using majority weighting, then
both this field and the test.set field are optional.}

\item{test.set}{The data test set so that weights could be gathered from the models. Not
needed if weights are given.}

\item{train.type}{The metric to be used in getting model weights. Not
needed if weights are given.}
}
\value{
An Ensemble Object
}
\description{
This function takes models and turns them into a combined Ensemble
}
\examples{
# prepare a binary classification dataset
iris <- PrepareIris()
# create the models
comp <- GetModelComparisons(iris[,1:4], iris[,5])
# use the models in the comparison to form a one model Ensemble
ensem <- Ensemble(comp$model.list, "majorityWeight", iris[,1:4], iris[,5])

}
