% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{GetBuildFlags}
\alias{GetBuildFlags}
\title{A helper function used by the GetModelComparison function. It takes the inputs and parses
it to see what models to build}
\usage{
GetBuildFlags(model.list)
}
\arguments{
\item{model.list}{The input parameters from GetModelComparison on what models to create.}
}
\value{
A vector of flags that contain what models to build.  To be used in BuildModels
function
}
\description{
A helper function used by the GetModelComparison function. It takes the inputs and parses
it to see what models to build
}
\examples{
# This function is used by GetModelComparison. It's usage is as follows:
build.flags <- GetBuildFlags(model.list)

}
\keyword{internal}
