% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembling.R
\name{GetFactorEqual}
\alias{GetFactorEqual}
\title{A quick helper function to flip the factors of 0's and 1's}
\usage{
GetFactorEqual(pred)
}
\arguments{
\item{pred}{The predictions from the model}
}
\value{
The predictions vector but in the correct format
}
\description{
A quick helper function to flip the factors of 0's and 1's
}
\examples{
# This function is used by GetWeightsFromTestSet It's usage is as follows:
# test.set <- GetFactorEqual(test.set)

}
\keyword{internal}
