% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembling.R
\name{GetModelWeights}
\alias{GetModelWeights}
\title{A helper function used by the Ensemble function. It will take the given inputs
and build the weights for each of the models.  This function mostly just handles different
cases and error checking and calls another function to get the weights data.}
\usage{
GetModelWeights(ensemble, weights, test.set, train.type)
}
\arguments{
\item{ensemble}{The training data to build the models.}

\item{weights}{Either a given vector of weights, or the value "none".}

\item{test.set}{The data test set so that weights could be gathered from the models. Not
needed if weights are given.}

\item{train.type}{The metric to be used in getting model weights. Not
needed if weights are given.}
}
\value{
A vector containg weights for the Ensemble's models.
}
\description{
A helper function used by the Ensemble function. It will take the given inputs
and build the weights for each of the models.  This function mostly just handles different
cases and error checking and calls another function to get the weights data.
}
\examples{
# This function is used by Ensemble. It's usage is as follows:
ensemble$weight.list = GetModelWeights(ensemble, weights, test.set, train.type)

}
\keyword{internal}
