% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackingHelpers.R
\name{GetPredictionsForStacking}
\alias{GetPredictionsForStacking}
\title{A helper function to get predictions added to a dataset.}
\usage{
GetPredictionsForStacking(models, x.values)
}
\arguments{
\item{models}{The models that will add their predictions to the DataFrame.}

\item{x.values}{The DataFrame that will be added on too and predicted on.}
}
\value{
The given DataFrame (x.values) with an additional column of predictions for each
model given.
}
\description{
A helper function to get predictions added to a dataset.
}
\examples{
iris <- PrepareIris()
x.values = iris[, 1:4]
# create the models
comp <- GetModelComparisons(x.values, iris[,5])
# get the new dataframe
df.for.stacking <- GetPredictionsForStacking(comp$model.list, x.values)

}
\keyword{internal}
