% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{GetPredType}
\alias{GetPredType}
\title{A helper function used by the predict function. Since the many types of models are not
uniform in how they are called, this function determines what values to pass into "type"
and then returns the predictions.}
\usage{
GetPredType(model, newdata)
}
\arguments{
\item{model}{The model to be used for prediction}

\item{newdata}{A dataframe for the model to predict on}
}
\value{
The predictions for the model on the given data, returned in dataframe format
}
\description{
A helper function used by the predict function. Since the many types of models are not
uniform in how they are called, this function determines what values to pass into "type"
and then returns the predictions.
}
\examples{
# This function is used by predict.ModelComparison. It's usage is as follows:
pred.basic[[i]] = GetPredType(model, newdata)

}
\keyword{internal}
