% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{GetSVMScale}
\alias{GetSVMScale}
\title{A helper function to ground SVM predictions}
\usage{
GetSVMScale(pred)
}
\arguments{
\item{pred}{The SVM predictions vector.}
}
\value{
The predictions, with only two factors.
}
\description{
SVM predictions are often greater than 1 or less than 0. We need to round them to make them
have only two factors
}
\examples{
# As used in GetCombinedPlot:
# pred <- GetSVMScale(pred)

}
\keyword{internal}
