% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{GetTrainingInfo}
\alias{GetTrainingInfo}
\title{A helper function used by the GetModelComparison function.}
\usage{
GetTrainingInfo(training.set, training.classes.input, validation,
  trctrl.given)
}
\arguments{
\item{training.set}{The training data to build the models.}

\item{training.classes.input}{The labels for the training data.}

\item{validation}{The type of model validation: cross-validation ("cv"), or a training
split (in the form XX/XX where XX is a two digit percent).}

\item{trctrl.given}{The trctrl given in GetModelComparison parameters for custom training.}
}
\value{
A vector containg the following for training the models. It is given in this order:
training.data, training.classes, trctrl
}
\description{
It takes the given inputs, data, and trctrl needed to build the models.
}
\examples{
# This function is used by GetModelComparison. It's usage is as follows:
training.info <- GetTrainingInfo(training.set, training.classes.input, validation, trctrl)

}
\keyword{internal}
