% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembling.R
\name{GetWeightsFromTestingSet}
\alias{GetWeightsFromTestingSet}
\title{A helper function used by the GetModelWeights function. It uses the given data to predict
and pull the metric from the data.}
\usage{
GetWeightsFromTestingSet(ensemble, df.train, test.set, train.type)
}
\arguments{
\item{ensemble}{The training data to build the models.}

\item{df.train}{The predictions of the model, given in DataFrame format.}

\item{test.set}{The data test set so that metrics can be calculated.}

\item{train.type}{The metric to be used in getting model weights. We pull this metric
from the confusion matrix.}
}
\value{
A vector containg weights for the Ensemble's models.
}
\description{
A helper function used by the GetModelWeights function. It uses the given data to predict
and pull the metric from the data.
}
\examples{
# This function is used by Ensemble. It's usage is as follows:
return(GetWeightsFromTestingSet(ensemble, weights, test.set, train.type))

}
\keyword{internal}
