% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{ModelComparison}
\alias{ModelComparison}
\title{A function to return a ModelComparison object.}
\usage{
ModelComparison(model.list, multi.class)
}
\arguments{
\item{model.list}{The training data to build the models.}

\item{multi.class}{A boolean of whether the models predict a multi-outcome response}
}
\value{
A ModelComparison object.
}
\description{
This function take pre-built models, and puts
them into a ModelComparison object.
}
\examples{
models <- list(model1, model2, model3)

names(models) <- c("NeuralNet", "K-NN", "SVM")

comp <- ModelComparison(models, F)

}
