% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plot.ModelComparison}
\alias{plot.ModelComparison}
\title{This function evalutates many different machine learning models and returns those models with comparison charts}
\usage{
\method{plot}{ModelComparison}(object, labels, training.data = "none",
  predictions = "empty", plot.type = "ROC", ...)
}
\arguments{
\item{plot.type}{A vector of characters that are the values seen in the plot
(examples include ROC, AUC, Accuracy, etc.)  Note: ROC cannot be plotted with other metrics.}

\item{trainingSet}{the dataset to be trained on}

\item{trainingClasses}{the labels of the training set}
}
\description{
This function evalutates many different machine learning models and returns those models with comparison charts
}
\examples{
plot()
}
