% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembling.R
\name{predict.Ensemble}
\alias{predict.Ensemble}
\title{Predicts values for a Ensemble object.}
\usage{
\method{predict}{Ensemble}(object, newdata, voting.type = "default", ...)
}
\arguments{
\item{object}{The Ensemble object whose models will be predicted on.}

\item{newdata}{A dataframe of new data that the models will use to predict.}

\item{voting.type}{The way that the Ensemble will vote - default is what was given
at creation of the object but it can be overriden by passing in "averageVote", "majorityVote"
or "majorityWeight."}
}
\value{
a list of predictions, voted on by all the models into one final prediction.
}
\description{
This involves calling the voting function specified
at Ensemble creation or can be overriden by parameters.
}
\examples{
# load the csv file for the dataset "titanic"
titanic <- PrepareNumericTitanic()
# create the ModelComparison object by passing in the training set and training labels
comp <- GetModelComparisons(titanic[, -1], titanic[, 1])
# use the models in the comparison to form a one model Ensemble
ensem <- Ensemble(comp$model.list, "majorityWeight", iris[,1:4], iris[,5])
# predict by passing in the new df for the object to predict on
pred.list <- predict(comp, titanic[, -1], voting.type="averageVote")

}
