% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{predict.ModelComparison}
\alias{predict.ModelComparison}
\title{Predicts values for a ModelComparison object.}
\usage{
\method{predict}{ModelComparison}(object, newdata, ...)
}
\arguments{
\item{object}{The ModelComparison object whose models will be predicted on.}

\item{newdata}{A dataframe of new data that the models will use to predict.}
}
\value{
A list of dataframes, where each dataframe is the predictions from the respective
           model for the two output classes.
}
\description{
Predicts values for a ModelComparison object.
}
\examples{
# load the csv file for the dataset "titanic"
titanic <- PrepareNumericTitanic()
# create the ModelComparison object by passing in the training set and training labels
comp <- GetModelComparisons(titanic[, -1], titanic[, 1])
# predict by passing in the new df for the object to predict on
pred.list <- predict(comp, titanic[, -1])

}
