% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackingHelpers.R
\name{PrepareIris}
\alias{PrepareIris}
\title{A quick helper function to turn the iris dataset into a 2 level dataset.}
\usage{
PrepareIris()
}
\value{
The Iris dataset, without the setosa rows
}
\description{
A quick helper function to turn the iris dataset into a 2 level dataset.
}
\examples{
iris <- PrepareIris()

}
\keyword{internal}
