% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackingHelpers.R
\name{PrepareNumericTitanic}
\alias{PrepareNumericTitanic}
\title{A quick helper function to turn the titanic dataset into a numeric only dataset.}
\usage{
PrepareNumericTitanic()
}
\value{
The Titanic dataset, but with only numeric columns.
}
\description{
A quick helper function to turn the titanic dataset into a numeric only dataset.
}
\examples{
titanic <- PrepareNumericTitanic()

}
\keyword{internal}
