% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{prepData}
\alias{prepData}
\title{A helper function used by the GetModelComparison function. If the data given is categorical
and not in one-hot encoding the model training will fail.  This will attempt to conver the
dataset into that format, or, to send an error if it fails to do so.}
\usage{
prepData(training.set)
}
\arguments{
\item{training.set}{the training data that is in the wrong format. This is the same dataframe
as in GetModelComparison parameters.}
}
\value{
The dataframe in one-hot encoding.
}
\description{
A helper function used by the GetModelComparison function. If the data given is categorical
and not in one-hot encoding the model training will fail.  This will attempt to conver the
dataset into that format, or, to send an error if it fails to do so.
}
\examples{
# This function is used by GetModelComparison. It's usage is as follows:
training.set = prepData(training.set)

}
\keyword{internal}
