% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembling.R
\name{StripPredictions}
\alias{StripPredictions}
\title{A function for extracting the prediction values from the models.  Each model returns the
prediction as a DataFrame and we need it as a list or vector.}
\usage{
StripPredictions(pred)
}
\arguments{
\item{pred}{A list containing a DataFrame of predictions.}
}
\value{
A vector containing the predictions.
}
\description{
A function for extracting the prediction values from the models.  Each model returns the
prediction as a DataFrame and we need it as a list or vector.
}
\examples{
# This function is used by predict.Ensemble. It's usage is as follows:
preds[[i]] <-  StripPredictions(predict(model, newdata = newdata, type="prob"))

}
\keyword{internal}
