% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{GetModelComparisons}
\alias{GetModelComparisons}
\title{This function evalutates many different machine learning models and returns those models with comparison charts}
\usage{
GetModelComparisons(training.set, training.classes.input,
  validation = "80/20", model.list = "fast", trctrl = "none")
}
\arguments{
\item{training.set}{the dataset to be trained on}

\item{trainingClasses}{the labels of the training set}
}
\description{
This function evalutates many different machine learning models and returns those models with comparison charts
}
\examples{
GetModelComparisons()
}
