% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelComparison-package.R, R/ModelGeneration.R
\docType{package}
\name{ModelComparison}
\alias{ModelComparison}
\alias{ModelComparison-package}
\title{ModelComparison: Visualization for Machine Learning Metrics}
\usage{
ModelComparison(model.list, multi.class)
}
\arguments{
\item{model.list}{The training data to build the models.}

\item{multi.class}{A boolean of whether the models predict a multi-outcome response}
}
\value{
A ModelComparison object.
}
\description{
Want to easily compare Machine Learning models?  Want to just summarize them in a pretty graph
without having to go to all the work?  This package consolidates this by allowing you to use
your pre-built models to make easy graphics or to build all those models for you.

This function take pre-built models, and puts
them into a ModelComparison object.
}
\examples{
models <- list(model1, model2, model3)

names(models) <- c("NeuralNet", "K-NN", "SVM")

comp <- ModelComparison(models, F)

}
